ai/pyproject.toml
[project]
name = "aldrax"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=24.10.0",
    "boto2>=0.1.0",
    "boto3>=1.36.2",
    "fastapi-limiter>=0.1.6",
    "fastapi>=0.115.6",
    "ollama>=0.4.6",
    "openai>=1.59.8",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "qdrant-client>=1.13.0",
    "redis>=5.2.1",
    "schedule>=1.2.2",
    "unicorn>=2.1.1",
    "uvicorn[standard]>=0.34.0",
    "yml>=0.0.1",
]
